combine the two rules into one
