+++ rules-tests/Php83/Rector/FuncCall/RemoveGetClassGetParentClassNoArgsRector/Fixture/fixture_get_class.php.inc
@@ -1,0 +1,27 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector\Fixture;
+
+class Fixture
+{
+    public function test()
+    {
+        echo get_class();
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector\Fixture;
+
+class Fixture
+{
+    public function test()
+    {
+        echo self::class;
+    }
+}
+
+?>
