+++ rules-tests/Php83/Rector/FuncCall/RemoveGetClassGetParentClassNoArgsRector/Fixture/skip_with_args.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector\Fixture;
+
+use Rector\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector;
+
+class SkipWithArgs
+{
+    public function test()
+    {
+        $obj = new RemoveGetClassGetParentClassNoArgsRector();
+        echo get_class($obj);
+        echo get_parent_class($obj);
+    }
+}
