+++ rules-tests/Php83/Rector/FuncCall/RemoveGetClassGetParentClassNoArgsRector/RemoveGetClassGetParentClassNoArgsRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetParentClassNoArgsRector;
+namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class RemoveGetParentClassNoArgsRectorTest extends AbstractRectorTestCase
+final class RemoveGetClassGetParentClassNoArgsRectorTest extends AbstractRectorTestCase
 {
     #[DataProvider('provideData')]
     public function test(string $filePath): void
