+++ rules-tests/Php83/Rector/FuncCall/RemoveGetClassGetParentClassNoArgsRector/config/configured_rule.php
@@ -3,11 +3,11 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Php83\Rector\FuncCall\RemoveGetClassNoArgsRector;
+use Rector\Php83\Rector\FuncCall\RemoveGetClassGetParentClassNoArgsRector;
 use Rector\ValueObject\PhpVersion;
 
 return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rule(RemoveGetClassNoArgsRector::class);
+    $rectorConfig->rule(RemoveGetClassGetParentClassNoArgsRector::class);
 
     $rectorConfig->phpVersion(PhpVersion::PHP_83);
 };
