--- rules-tests/Php83/Rector/FuncCall/RemoveGetClassNoArgsRector/RemoveGetClassNoArgsRectorTest.php
@@ -1,28 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetClassNoArgsRector;
-
-use Iterator;
-use PHPUnit\Framework\Attributes\DataProvider;
-use Rector\Testing\PHPUnit\AbstractRectorTestCase;
-
-final class RemoveGetClassNoArgsRectorTest extends AbstractRectorTestCase
-{
-    #[DataProvider('provideData')]
-    public function test(string $filePath): void
-    {
-        $this->doTestFile($filePath);
-    }
-
-    public static function provideData(): Iterator
-    {
-        return self::yieldFilesFromDirectory(__DIR__ . '/Fixture');
-    }
-
-    public function provideConfigFilePath(): string
-    {
-        return __DIR__ . '/config/configured_rule.php';
-    }
-}
