--- rules-tests/Php83/Rector/FuncCall/RemoveGetParentClassNoArgsRector/Fixture/skip_different_func_call.php.inc
@@ -1,11 +1,0 @@
-<?php
-
-namespace Rector\Tests\Php83\Rector\FuncCall\RemoveGetParentClassNoArgsRector\Fixture;
-
-class SkipDifferentFuncCall
-{
-    public function run($a, $b)
-    {
-        return str_contains($a, $b);
-    }
-}
